image/requirements.txt
rembg-aws-lambda==0.2.0
pillow<=11.1.0
onnxruntime<=1.20.1